trie scenarios
